Fixed #1172 - 2.0: Database API Changes (DB008) (#1177)
- compact
- contains
- count
- other minor updates.

+++ shared/src/main/java/com/couchbase/lite/ReadOnlyDictionary.java
@@ -50,7 +50,7 @@ public class ReadOnlyDictionary implements ReadOnlyDictionaryInterface, FleeceEn
      */
     @Override
     public int count() {
-        return flDict != null ? flDict.count() : 0;
+        return flDict != null ? (int) flDict.count() : 0;
     }
 
     /**
